Add reimplement to use WebCryptop and fallback to js
