+++ src/aes/aes_base.ts
@@ -1,0 +1,4 @@
+export abstract class AESBase {
+  public abstract async encrypt(m: Uint8Array): Promise<Uint8Array>;
+  public abstract async decrypt(m: Uint8Array): Promise<Uint8Array>;
+}
