+++ src/aes/common.ts
@@ -1,0 +1,12 @@
+export interface BlockCiperConfig {
+  mode?: "ctr" | "cfb" | "ofb" | "ecb" | "cbc";
+  padding?: "pkcs5";
+  iv?: Uint8Array | string;
+}
+
+type AESBlockMode = "cbc" | "ecb";
+
+export interface AESOption {
+  mode: AESBlockMode;
+  iv: string | Uint8Array;
+}
